macro refactor (no logic change)
