+++ core/shared/src/main/scala-3/org/scalamock/clazz/MakerUtils.scala
@@ -1,0 +1,120 @@
+package org.scalamock.clazz
+
+import org.scalamock.stubs.Stub
+import org.scalamock.util.Defaultable
+
+import scala.annotation.{experimental, tailrec}
+import scala.quoted.{Expr, Quotes, Type}
+
+private[scalamock]
+class MakerUtils(using override val quotes: Quotes) extends Utils:
+  import quotes.reflect.*
+
+  def parentsOf[T: Type]: List[TypeTree | Term] =
+    val tpe = TypeRepr.of[T]
+    val isTrait = tpe.dealias.typeSymbol.flags.is(Flags.Trait)
+    val isJavaClass = tpe.classSymbol.exists(sym => sym.flags.is(Flags.JavaDefined) && !sym.flags.is(Flags.Trait) && !sym.flags.is(Flags.Abstract))
+
+    if (isJavaClass)
+      report.errorAndAbort("Can't mock a java class due to https://github.com/lampepfl/dotty/issues/18694. Extend it manually with scala and then mock")
+
+    def asParent(tree: TypeTree): TypeTree | Term =
+      val constructorTypes = tree.tpe.typeSymbol.primaryConstructor
+        .paramSymss.flatten.filter(_.isType).map(_.name)
+
+      val constructorFields = tree.tpe.dealias.typeSymbol.primaryConstructor
+        .paramSymss.filterNot(_.exists(_.isType))
+
+      if tree.tpe.typeArgs.length < constructorTypes.length then
+        report.errorAndAbort("Not all types are applied")
+
+      if constructorFields.forall(_.isEmpty) then
+        tree
+      else
+        val con = Select(
+          New(TypeIdent(tree.tpe.typeSymbol)),
+          tree.tpe.typeSymbol.primaryConstructor
+        ).appliedToTypes(tree.tpe.typeArgs)
+
+        val typeNames = {
+          @tailrec
+          def collectTypes(tpe: TypeRepr, acc: Map[String, TypeRepr]): Map[String, TypeRepr] =
+            tpe match
+              case MethodType(names, types, res) => collectTypes(res, acc ++ names.zip(types))
+              case tpe => acc
+
+          collectTypes(con.tpe.widenTermRefByName, Map.empty)
+            .view.mapValues {
+              case n: ByNameType => n.widenByName
+              case n => n
+            }
+            .toMap
+        }
+
+        con.appliedToArgss(
+          constructorFields
+            .map(_.map { sym => typeNames(sym.name).asType })
+            .map(_.map { case '[t] => '{ ${ Expr.summon[Defaultable[t]].get }.default }.asTerm })
+        )
+    end asParent
+
+    if isTrait then
+      List(TypeTree.of[Object], asParent(TypeTree.of[T]), TypeTree.of[scala.reflect.Selectable])
+    else
+      List(asParent(TypeTree.of[T]), TypeTree.of[scala.reflect.Selectable])
+  end parentsOf
+
+  @experimental
+  def searchTermWithMethod(term: Term, argTypes: List[TypeRepr], appliedTypes: List[TypeTree] = Nil): StubWithMethod =
+    term match
+      case Select(mock, methodName) =>
+        val tpe =
+          mock.tpe.widenTermRefByName match
+            case AppliedType(tycon, List(typ)) if newApi && tycon =:= TypeRepr.of[Stub] =>
+              typ
+            case tpe if newApi =>
+              report.errorAndAbort(s"This syntax is only available on Stub[T]")
+            case other =>
+              other
+
+        val method = MockableDefinitions.find(tpe, methodName, argTypes, appliedTypes.map(_.tpe))
+        StubWithMethod(mock, method)
+
+      case Block(_, Inlined(_, _, Inlined(_, _, term: Select))) =>
+        searchTermWithMethod(term, argTypes)
+
+      case Inlined(_, _, term) =>
+        searchTermWithMethod(term, argTypes)
+
+      case Block(stats@List(_: ValDef), term) => // var m = mock[T]
+        val StubWithMethod(underlying, method) = searchTermWithMethod(term, argTypes)
+        StubWithMethod(Block(stats, underlying), method)
+
+      case Block(List(DefDef(_, _, _, Some(term))), _) =>
+        searchTermWithMethod(term, argTypes)
+
+      case Typed(term, teps) =>
+        searchTermWithMethod(term, argTypes)
+
+      case Lambda(_, term) =>
+        searchTermWithMethod(term, argTypes)
+
+      case Apply(term, _) =>
+        searchTermWithMethod(term, argTypes)
+
+      case TypeApply(term, types) =>
+        searchTermWithMethod(term, argTypes, types)
+
+      case Ident(fun) if !newApi =>
+        report.errorAndAbort(
+          s"please declare '$fun' as MockFunctionX or StubFunctionX (e.g val $fun: MockFunction1[X, R] = ... if it has 1 parameter)"
+        )
+      case _ =>
+        report.errorAndAbort(
+          s"""
+              This syntax is only available on Stub[T].
+              Current structure is not recognised: ${term.show}
+              If you think this is a bug - please open a ticket at https://github.com/ScalaMock/ScalaMock/issues"
+          """
+        )
+  end searchTermWithMethod
